feat: Implement comDevelopment mode for accessing hidden settings
By default PrusaSlicer has comSimple, comAdvanced and comExpert for grouping slicing settings and tools. The new comDevelopment is designed to grant access to core features and settings that should not be changed by customers. For now they can only be access by defining the APIUMSLICER_DEV_MODE in the apium_config.hpp. Releases for customers shall not be compiled with APIUMSLICER_DEV_MODE enabled.